+++ src/main/java/com/texnoera/socialmedia/service/abstracts/UserService.java
@@ -12,12 +12,12 @@ import java.util.List;
 public interface UserService extends UserDetailsService {
 
     List<UserResponse> getAll();
-    UserResponse getResponseById(Long id);
+    UserResponse getResponseById(Integer id);
     UserResponse getByEmail(String email);
-    List<UserFollowingResponse> getUserFollowing(Long userId);
-    boolean isFollowing(Long userId, Long followingId);
-    User getById(Long id);
+    List<UserFollowingResponse> getUserFollowing(Integer userId);
+    boolean isFollowing(Integer userId, Integer followingId);
+    User getById(Integer id);
     UserResponse add(UserAddRequest userAddRequest);
-    void delete(Long id);
-    UserResponse update(Long id, UserUpdateRequest userUpdateRequest);
+    void delete(Integer id);
+    UserResponse update(Integer id, UserUpdateRequest userUpdateRequest);
 }
